// src/app/shared/model/vendor-atpm-mapping-obj.model.ts
export class VendorAtpmMappingObj {
    VendorAtpmMappingId: number;
    VendorId: number;
    VendorAtpmId: number;
    VendorCode: string;
    VendorName: string;
    VendorLegalAddr: string;
    VendorAtpmCode: string;
    VendorAtpmName: string;
    VendorAtpmLegalAddr: string;

    constructor() { this.VendorAtpmMappingId = 0; }
}  